requirement.txt
Django
djangorestframework
djangorestframework-simplejwt
requests
django-cors-header
pillow